Another restructure & Wit.ai WORKS!
Wit.ai is now implemented, conversation will flow nicely from the Messenger user to Wit.ai and back to the user. Even got the sessions working nicely (at least it seems, we'll know once we need context), and sucessfully got mark_seen and typing_on and typing_off to work to engage the user while their message is being processed. That's a win!