We're live boiz!
This is it, the bot is talking to Wit.ai with no issues, things are running smoothly. What more could you want? Oh.. emotions? Right... Well, that's probably where we'll go next, give the bot some personality, start adding more commands and things for it to do etc. Build a good base to have a useful conversation with it. Make it smart, make it personal. Right now it just says hello to greetings and finds all shows that are listed in sick rage when asked to do so. That's it! Tada :D