IT HAS FEELINGS!!! I mean... he has feelings..
The bot can now feel very basic happiness, sadness and neutrality. Happiness is the scale and it's affected by how the user feels, if it's positive or negative, it will drive the happiness up/down. The happiness of the bot is reflected only in its greetings, the happiness will be reflected in more later down the road. Winning for now though!